+++ govc/device/serial/connect.go
@@ -19,9 +19,12 @@ package serial
 import (
 	"context"
 	"flag"
+	"fmt"
+	"path"
 
 	"github.com/vmware/govmomi/govc/cli"
 	"github.com/vmware/govmomi/govc/flags"
+	"github.com/vmware/govmomi/vim25/mo"
 )
 
 type connect struct {
@@ -45,13 +48,25 @@ func (cmd *connect) Register(ctx context.Context, f *flag.FlagSet) {
 	f.BoolVar(&cmd.client, "client", false, "Use client direction")
 }
 
+func (cmd *connect) Usage() string {
+	return "URI"
+}
+
 func (cmd *connect) Description() string {
 	return `Connect service URI to serial port.
+
+If "-" is given as URI, connects file backed device with file name of
+device name + .log suffix in the VM Config.Files.LogDirectory.
+
+Defaults to the first serial port if no DEVICE is given.
 
 Examples:
   govc device.ls | grep serialport-
   govc device.serial.connect -vm $vm -device serialport-8000 telnet://:33233
-  govc device.info -vm $vm serialport-*`
+  govc device.info -vm $vm serialport-*
+  govc device.serial.connect -vm $vm "[datastore1] $vm/console.log"
+  govc device.serial.connect -vm $vm -
+  govc datastore.tail -f $vm/serialport-8000.log`
 }
 
 func (cmd *connect) Process(ctx context.Context) error {
@@ -62,6 +77,10 @@ func (cmd *connect) Process(ctx context.Context) error {
 }
 
 func (cmd *connect) Run(ctx context.Context, f *flag.FlagSet) error {
+	if f.NArg() != 1 {
+		return flag.ErrHelp
+	}
+
 	vm, err := cmd.VirtualMachine()
 	if err != nil {
 		return err
@@ -81,5 +100,17 @@ func (cmd *connect) Run(ctx context.Context, f *flag.FlagSet) error {
 		return err
 	}
 
-	return vm.EditDevice(ctx, devices.ConnectSerialPort(d, f.Arg(0), cmd.client, cmd.proxy))
+	uri := f.Arg(0)
+
+	if uri == "-" {
+		var mvm mo.VirtualMachine
+		err = vm.Properties(ctx, vm.Reference(), []string{"config.files.logDirectory"}, &mvm)
+		if err != nil {
+			return err
+		}
+
+		uri = path.Join(mvm.Config.Files.LogDirectory, fmt.Sprintf("%s.log", devices.Name(d)))
+	}
+
+	return vm.EditDevice(ctx, devices.ConnectSerialPort(d, uri, cmd.client, cmd.proxy))
 }