+++ govc/test/boot_order_test.sh
@@ -53,16 +53,15 @@ govc vm.power -off $id
 
 govc device.cdrom.eject -vm $id
 
-host_ip=$(echo $vnc | awk -F@ '{print $2}' | awk -F: '{print $1}')
-serial_port=33233
 govc device.serial.add -vm $id > /dev/null
-govc device.serial.connect -vm $id $uri telnet://:$serial_port
+govc device.serial.connect -vm $id -
 
 echo "booting from network, will timeout then boot from disk..."
 govc vm.power -on $id
 
-# capture serial console
-echo | nc $host_ip $serial_port 2>/dev/null &
+# serial console log
+device=$(govc device.ls -vm "$id" | grep serialport- | awk '{print $1}')
+govc datastore.tail -f "$id/$device.log" &
 
 ip=$(govc vm.ip $id)